Fixed second task

+++ task_2/Main.java
@@ -5,8 +5,8 @@ public class Main {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
 
-        double a = in.nextInt();
-        double b = in.nextInt();
+        double a = in.nextDouble();
+        double b = in.nextDouble();
           
         double d = 
             (Math.max(2 * a, b) - 10 * Math.sqrt(Math.min(a, 4 + b)) + 4.2 * Math.min(a, 4 + b))
